feat: added thumbnail generation logi and wrote test for Fetch function
